ingestion/requirements.txt
httpx==0.27.0
tenacity==8.2.3
sqlalchemy==2.0.29
psycopg[binary]==3.1.18
pydantic==2.7.1
python-dotenv==1.0.1
pytest==8.2.2
black==24.8.0
ruff==0.5.7
mypy==1.10.0
PyYAML==6.0.2
